backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
sentence-transformers>=2.3.0
pdfplumber==0.10.3
python-docx==1.1.0
python-multipart==0.0.6
pydantic==2.5.0
numpy
torch
